YT-web-application/requirements.txt
Flask==3.0.0
yt-dlp>=2025.10.22
python-dotenv==1.0.0
Werkzeug==3.0.1
